+++ packages/core-cairo/src/external-trait.ts
@@ -4,7 +4,8 @@ export const externalTrait: BaseImplementedTrait = {
   name: 'ExternalImpl',
   of: 'ExternalTrait',
   tags: [
-    '#[generate_trait]',
-    '#[external(v0)]'
+    'generate_trait',
+    'abi(per_item)',
   ],
+  perItemTag: 'external(v0)',
 }